Add useState to Each Element to View It In Preview
The Hook useState is added for each section (Personal info, contact infor, etc.) to get the data out of the input using the state variables, then the preview panel is updated live, or once the button is pressed. The state is then cleared. The preview panel is updated using another useState hook, which makes an object of all elements of the same section, for example (workList object contains work1 which has the position, company, etc.). This state variable is used to update the screen using the state function once the button is pressed.